+++ react/context/index.tsx
@@ -1,1 +1,1 @@
-export * from './ShippingOptionContext'
+export * from './DeliveryPromiseContext'